finishing of next step panels
