+++ src/App.css
@@ -3,11 +3,11 @@
 /* prefixed by https://autoprefixer.github.io (PostCSS: v7.0.26, autoprefixer: v9.7.3) */
 
 .App {
+    min-height: 90vh;
     text-align: center;
-    min-width: 900px;
+    min-width: 1200px;
     background-color: white;
     color: black;
-    height: 90vh;
     margin-right: 5%;
     margin-left: 5%;
 }
@@ -41,16 +41,17 @@
 
 }
 img{
-    height: 70%;
+    height: 10vh;
     -ms-flex-item-align: center;
         -ms-grid-row-align: center;
         align-self: center;
     margin-left: 1%;
+    margin: 1%;
 }
 
 .Body{
+    max-height: 90vh;
     color: black; 
-    height: 90%;
     display: -webkit-box;
     display: -ms-flexbox;
     display: flex;  
@@ -68,8 +69,9 @@ img{
     
 }
 .filter-container > div {
-    border-bottom-style: groove;
+    border-bottom-style: solid;
     border-bottom-width: thin;
+    border-bottom-color: #707070;
 }
 .chart-panel {
     width: 70%;
@@ -78,4 +80,42 @@ img{
     overflow-y: auto;
     background-color: white;
     color: black;
+}
+h1{
+    font-size: 80px;
+    font-weight: normal;
+    font: arial;
+    text-align: center;
+}
+.Boxes{
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    flex-wrap: wrap;
+    font-size: 28px;
+    margin-left: 13%;
+    margin-right: 13%;
+    text-align: center;
+}
+.box{
+    margin-top: 4%;
+    margin-bottom: 3.5%;
+    padding-top: 2%;
+    padding-bottom: 2%;
+    width: 45%;
+    border-style: solid;
+    border-width: normal;
+    border-color: 	#707070;
+}
+.learnMore{
+    margin-left: 10%;
+    margin-right: 10%;
+    margin-top: 6%;
+    text-align: center;
+    font-size: 28px;
+}
+footer{
+    background-color: #707070;
+    margin: 0em;
+    display: block;
 }
