+++ src/App.js
@@ -11,7 +11,7 @@ import EconDisChart from "./components/Charts/EconDisChart";
 import DisabilityChart from "./components/Charts/DisabilityChart";
 import ELLChart from "./components/Charts/ELLChart";
 import CourseEnrollmentChart from "./components/Charts/CourseEnrollment"
-import NextStepsPanel from "./components/NextStepsPanel";
+import NextStepsPanel from "./components/NextSteps/NextStepsPanel";
 
 import schoolData from "./data/data-new";
 
@@ -70,29 +70,32 @@ class App extends React.Component {
         const charts = this.createChartsFromFilterState();
 
         return (
-            <div className="App">
-                <div className="Header">
-                    <img src={logo} alt="Logo" />
-                    <h3> Digital Equity Initiative </h3>
-                </div>
-                <div className="Body">
-                    <div className="filter-panel">
-                        <FilterPanel
-                            data={this.state.schoolData}
-                            selectedFilters={this.state.selectedFilters}
-                            selectedYears={this.state.selectedYearsMap}
-                            onSchoolFilterChange={(opts, actionMeta) =>
-                                this.schoolFilterChangeHandler(opts, actionMeta)
-                            }
-                            onGraphSelectionChange={
-                                this.graphSelectionChangeHandler
-                            }
-                            onYearSelectionChange={this.yearSelectionChangeHandler}
-                        />
+            <div>
+                <div className="App">
+                    <div className="Header">
+                        <img src={logo} alt="Logo" />
+                        <h3> Digital Equity Initiative </h3>
                     </div>
-                    <div className="chart-panel">{charts}</div>
+                    <div className="Body">
+                        <div className="filter-panel">
+                            <FilterPanel
+                                data={this.state.schoolData}
+                                selectedFilters={this.state.selectedFilters}
+                                selectedYears={this.state.selectedYearsMap}
+                                onSchoolFilterChange={(opts, actionMeta) =>
+                                    this.schoolFilterChangeHandler(opts, actionMeta)
+                                }
+                                onGraphSelectionChange={
+                                    this.graphSelectionChangeHandler
+                                }
+                                onYearSelectionChange={this.yearSelectionChangeHandler}
+                            />
+                        </div>
+                        <div className="chart-panel">{charts}</div>
+                    </div>
                 </div>
                 <NextStepsPanel />
+                <footer>Contact Info</footer>
             </div>
         );
     }
