+++ src/components/NextSteps/NextStepsPanel.js
@@ -1,0 +1,25 @@
+import React from "react";
+import "../../App.css";
+
+function NextStepsPanel() {
+
+    return (
+        <div>
+            <h1>Next Steps</h1>
+            <div className="Boxes">
+                <div className="box"> Closing the Digital Equity Gap</div>
+                <div className="box"> The Digital Equity Walk </div>
+                <div className="box">The Walk in Action</div>
+                <div className="box">Take the Virtual Walk</div>
+            </div>
+            <div className="learnMore">
+                Want to learn more about the Digital Equity Walk or about
+                opportunities to host a walk? Visit the Digital Equity Walk
+                in Action or contact Jackney Prioly Joseph.
+            </div>
+
+        </div>
+    );
+}
+
+export default NextStepsPanel;
